extract campaign definition routes
