+++ src/api/routers/waves.py
@@ -32,10 +32,9 @@ from src.api.routers.wave_request_models import (
 )
 from src.api.routers.wave_campaign_definition_http import (
     get_campaign_definition_response,
-    list_campaign_definitions_response,
-    put_campaign_definition_response,
-    retire_campaign_definition_response,
-    supersede_campaign_definition_response,
+)
+from src.api.routers.wave_campaign_definition_routes import (
+    router as campaign_definition_router,
 )
 from src.api.routers.wave_campaign_read_model_routes import (
     router as campaign_read_model_router,
@@ -43,12 +42,9 @@ from src.api.routers.wave_campaign_read_model_routes import (
 from src.api.routers.wave_route_parameters import (
     CampaignActiveOnQuery,
     CampaignAssignmentTaskRefPath,
-    CampaignDefinitionAsOfDateQuery,
     CampaignEvidenceLimitQuery,
     CampaignEvidenceOffsetQuery,
-    CampaignDefinitionFilterIdQuery,
     CampaignDefinitionIdPath,
-    CampaignDefinitionStatusQuery,
     CampaignDefinitionVersionPath,
     CampaignIncludeLaunchPackageQuery,
     CampaignLaunchActorIdOptionalQuery,
@@ -57,8 +53,6 @@ from src.api.routers.wave_route_parameters import (
     CampaignLaunchHistoryLimitQuery,
     CampaignLaunchHistoryOffsetQuery,
     CampaignLaunchRequestedAsOfDateQuery,
-    CampaignReadModelLimitQuery,
-    CampaignReadModelOffsetQuery,
     WaveCorrelationIdHeader,
     WaveCreateIdempotencyKeyHeader,
     WaveIdPath,
@@ -71,10 +65,6 @@ from src.api.routers.wave_campaign_models import (
     DpmBulkReviewCampaignDefinitionAssignmentTaskTransitionRequest,
     DpmBulkReviewCampaignDefinitionLaunchRequest,
     DpmBulkReviewCampaignDefinitionMakerCheckerControlRequest,
-    DpmBulkReviewCampaignDefinitionPage,
-    DpmBulkReviewCampaignDefinitionRequest,
-    DpmBulkReviewCampaignDefinitionRetirementRequest,
-    DpmBulkReviewCampaignDefinitionSupersessionRequest,
 )
 from src.api.routers.wave_campaign_action_http import (
     list_campaign_definition_approval_decisions_response,
@@ -151,118 +141,7 @@ from src.infrastructure.advise_authority import (
 router = APIRouter(prefix="/rebalance/waves", tags=["lotus-manage Rebalance Waves"])
 logger = logging.getLogger(__name__)
 
-@router.put(
-    "/campaign-definitions/{campaign_id}/versions/{campaign_version}",
-    response_model=DpmBulkReviewCampaignDefinition,
-    status_code=status.HTTP_200_OK,
-    summary="Persist bulk-review campaign definition",
-    description=(
-        "Persists an immutable Manage-owned `BulkReviewCampaignDefinition:v1` over a bounded, "
-        "source-backed candidate portfolio set. This endpoint does not discover the global book, "
-        "own source facts, run maker-checker workflow, expose downstream UI, or claim OMS "
-        "execution."
-    ),
-)
-def put_bulk_review_campaign_definition(
-    campaign_id: CampaignDefinitionIdPath,
-    campaign_version: CampaignDefinitionVersionPath,
-    request: DpmBulkReviewCampaignDefinitionRequest,
-    repository: DpmBulkReviewCampaignDefinitionRepository = Depends(
-        get_campaign_definition_repository
-    ),
-) -> DpmBulkReviewCampaignDefinition:
-    return put_campaign_definition_response(
-        campaign_id=campaign_id,
-        campaign_version=campaign_version,
-        request=request,
-        repository=repository,
-    )
-
-
-@router.get(
-    "/campaign-definitions",
-    response_model=DpmBulkReviewCampaignDefinitionPage,
-    status_code=status.HTTP_200_OK,
-    summary="List bulk-review campaign definitions",
-    description="Lists immutable Manage-owned bulk-review campaign definitions.",
-)
-def list_bulk_review_campaign_definitions(
-    campaign_id: CampaignDefinitionFilterIdQuery = None,
-    campaign_status: CampaignDefinitionStatusQuery = None,
-    as_of_date: CampaignDefinitionAsOfDateQuery = None,
-    limit: CampaignReadModelLimitQuery = 50,
-    offset: CampaignReadModelOffsetQuery = 0,
-    repository: DpmBulkReviewCampaignDefinitionRepository = Depends(
-        get_campaign_definition_repository
-    ),
-) -> DpmBulkReviewCampaignDefinitionPage:
-    return list_campaign_definitions_response(
-        campaign_id=campaign_id,
-        campaign_status=campaign_status,
-        as_of_date=as_of_date,
-        limit=limit,
-        offset=offset,
-        repository=repository,
-    )
-
-
-@router.post(
-    "/campaign-definitions/{campaign_id}/versions/{campaign_version}/retire",
-    response_model=DpmBulkReviewCampaignDefinition,
-    status_code=status.HTTP_200_OK,
-    summary="Retire bulk-review campaign definition",
-    description=(
-        "Retires a persisted Manage-owned `BulkReviewCampaignDefinition:v1` so it remains "
-        "auditable but can no longer be used for new `BULK_REVIEW_CAMPAIGN` preview/create "
-        "requests. This lifecycle action does not change the source-backed candidate set, "
-        "discover a global portfolio universe, run maker-checker workflow, or claim OMS execution."
-    ),
-)
-def retire_bulk_review_campaign_definition(
-    campaign_id: CampaignDefinitionIdPath,
-    campaign_version: CampaignDefinitionVersionPath,
-    request: DpmBulkReviewCampaignDefinitionRetirementRequest,
-    repository: DpmBulkReviewCampaignDefinitionRepository = Depends(
-        get_campaign_definition_repository
-    ),
-) -> DpmBulkReviewCampaignDefinition:
-    return retire_campaign_definition_response(
-        campaign_id=campaign_id,
-        campaign_version=campaign_version,
-        request=request,
-        repository=repository,
-    )
-
-
-@router.post(
-    "/campaign-definitions/{campaign_id}/versions/{campaign_version}/supersede",
-    response_model=DpmBulkReviewCampaignDefinition,
-    status_code=status.HTTP_200_OK,
-    summary="Supersede bulk-review campaign definition",
-    description=(
-        "Supersedes a persisted Manage-owned `BulkReviewCampaignDefinition:v1` with an already "
-        "persisted ACTIVE replacement version for the same campaign id. Superseded definitions "
-        "remain auditable but cannot be used for new `BULK_REVIEW_CAMPAIGN` preview/create "
-        "requests. This lifecycle action does not discover the global portfolio universe, "
-        "recalculate source facts, run maker-checker workflow, or claim OMS execution."
-    ),
-)
-def supersede_bulk_review_campaign_definition(
-    campaign_id: CampaignDefinitionIdPath,
-    campaign_version: CampaignDefinitionVersionPath,
-    request: DpmBulkReviewCampaignDefinitionSupersessionRequest,
-    repository: DpmBulkReviewCampaignDefinitionRepository = Depends(
-        get_campaign_definition_repository
-    ),
-) -> DpmBulkReviewCampaignDefinition:
-    return supersede_campaign_definition_response(
-        campaign_id=campaign_id,
-        campaign_version=campaign_version,
-        request=request,
-        repository=repository,
-    )
-
-
+router.include_router(campaign_definition_router)
 router.include_router(campaign_read_model_router)
